CIS 155 Project V4
This is the final version of my site for CIS 155. I primarily cleaned up some of my code and added a color transition to the hamburger menu when in mobile view. I also added two new media elements to the showcase page towards the bottom. In desktop view they will be side by side, and in mobile the columns collapse to one view.